Move validateAndroidArtifact to separate file
